Basic Framework

+++ src/PID.cpp
@@ -1,6 +1,6 @@
 #include "PID.h"
 
-uint8_t speed = 0;  // Controls the motor speed to balance the pendulum;
+int16_t speed = 0;  // Controls the motor speed to balance the pendulum;
 
 int16_t PIDController(int8_t angle)
 {
